Remove spkg scheme
Installing system packages depends very much on the particular operating
system. I could not find a good solution to enable that cross-platform,
so I'm removing that feature.

I'm not opposed to a spkg scheme, but it must be simple to use and not lay
too much burden on the manifest authors. In the meantime, use a path scheme
and run the install command in the action.